requirements.txt
# requirements.txt for LLM Analysis Quiz API
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
httpx==0.25.1
pydantic==2.5.0
python-dotenv==1.0.0
